bunch of redis stuff
